backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
opencv-python
Pillow
requests
ultralytics
numpy
psycopg2-binary
Flask-Cors
lucid
werkzeug